# openapispec.json
{
  "servers": [
    {
      "url": "https://0.0.0.0:3000/api/rpc"
    }
  ],
  "info": {
    "title": "API Reference",
    "version": "0.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/incidents": {
      "get": {
        "operationId": "incidents.list",
        "summary": "List incidents",
        "description": "Retrieve a list of incidents with optional filtering.",
        "tags": [
          "Incident Management"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 50
            },
            "allowEmptyValue": true,
            "allowReserved": true
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 0
            },
            "allowEmptyValue": true,
            "allowReserved": true
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "enum": [
                "open",
                "resolved",
                "all"
              ],
              "type": "string",
              "default": "all"
            },
            "allowEmptyValue": true,
            "allowReserved": true
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "allowEmptyValue": true,
            "allowReserved": true
          },
          {
            "name": "severity",
            "in": "query",
            "schema": {
              "enum": [
                "minor",
                "major",
                "critical",
                "maintenance"
              ],
              "type": "string"
            },
            "allowEmptyValue": true,
            "allowReserved": true
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "enum": [
                "manual",
                "automatic"
              ],
              "type": "string"
            },
            "allowEmptyValue": true,
            "allowReserved": true
          },
          {
            "name": "monitorId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "allowEmptyValue": true,
            "allowReserved": true
          },
          {
            "name": "statusPageId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "allowEmptyValue": true,
            "allowReserved": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "incidents.create",
        "summary": "Create incident",
        "description": "Create a new incident.",
        "tags": [
          "Incident Management"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "severity": {
                    "enum": [
                      "minor",
                      "major",
                      "critical",
                      "maintenance"
                    ],
                    "type": "string"
                  },
                  "monitorIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": []
                  },
                  "statusPageIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": []
                  },
                  "startedAt": {
                    "type": "string",
                    "format": "date-time",
                    "x-native-type": "date"
                  },
                  "plannedEndAt": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "date-time",
                        "x-native-type": "date"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "endedAt": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "date-time",
                        "x-native-type": "date"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "title",
                  "severity"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/incidents/{id}": {
      "get": {
        "operationId": "incidents.get",
        "summary": "Get incident",
        "description": "Retrieve details of a specific incident.",
        "tags": [
          "Incident Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "incidents.update",
        "summary": "Update incident",
        "description": "Update incident metadata, timeline, monitors, and publishing.",
        "tags": [
          "Incident Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "severity": {
                    "enum": [
                      "minor",
                      "major",
                      "critical",
                      "maintenance"
                    ],
                    "type": "string"
                  },
                  "startedAt": {
                    "type": "string",
                    "format": "date-time",
                    "x-native-type": "date"
                  },
                  "plannedEndAt": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "date-time",
                        "x-native-type": "date"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "endedAt": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "date-time",
                        "x-native-type": "date"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "monitorIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": []
                  },
                  "statusPageIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": []
                  }
                },
                "required": [
                  "title",
                  "severity",
                  "startedAt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "incidents.delete",
        "summary": "Delete incident",
        "description": "Delete a specific incident by ID.",
        "tags": [
          "Incident Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/maintenance": {
      "get": {
        "operationId": "maintenance.list",
        "summary": "List maintenance windows",
        "tags": [
          "Status Page Management"
        ],
        "parameters": [
          {
            "name": "statusPageId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "allowEmptyValue": true,
            "allowReserved": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "maintenance.create",
        "summary": "Create maintenance",
        "tags": [
          "Status Page Management"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "statusPageId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "startAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "status": {
                    "enum": [
                      "scheduled",
                      "in_progress",
                      "completed"
                    ],
                    "type": "string"
                  },
                  "monitorIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "statusPageId",
                  "title",
                  "startAt",
                  "endAt",
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/maintenance/{maintenanceId}": {
      "get": {
        "operationId": "maintenance.get",
        "summary": "Get maintenance",
        "tags": [
          "Status Page Management"
        ],
        "parameters": [
          {
            "name": "maintenanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "maintenance.update",
        "summary": "Update maintenance",
        "tags": [
          "Status Page Management"
        ],
        "parameters": [
          {
            "name": "maintenanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "startAt",
                  "endAt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "maintenance.delete",
        "summary": "Delete maintenance",
        "tags": [
          "Status Page Management"
        ],
        "parameters": [
          {
            "name": "maintenanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {},
                    {
                      "not": {}
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}